fix: better event management

+++ src/types/data.rs
@@ -272,7 +272,7 @@ pub struct Amount {
     /// Value of the transaction in atomic units represented as an
     /// arbitrary-sized signed integer.  For example, 1 BTC would be represented
     /// by a value of 100000000.
-    pub value: u128,
+    pub value: u64,
 
     pub currency: Currency,
     /* Rosetta Spec also optionally provides:
